+++ udata_front/theme/gouvfr/assets/js/components/AdminMembershipRequest/AdminMembershipRequest.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Controls, Description, Meta } from '@storybook/blocks';
 
-import * as MembershipRequestStories from './MembershipRequest.stories';
+import * as MembershipRequestStories from './AdminMembershipRequest.stories';
 
 <Meta of={MembershipRequestStories} />
